+++ app/services/emails/Email.scala
@@ -1,17 +1,10 @@
 package services.emails
 
 import models.report.Report
-import models.EmailValidation
-import play.api.i18n.Lang
-import play.api.i18n.MessagesApi
-import play.api.i18n.MessagesImpl
-import play.api.i18n.MessagesProvider
 import play.api.libs.mailer.Attachment
 import services.AttachmentService
 import utils.EmailAddress
 import utils.FrontRoute
-
-import java.util.Locale
 
 trait Email {
   val recipients: Seq[EmailAddress]
@@ -31,30 +24,4 @@ trait ProFilteredEmailMultipleReport extends ProFilteredEmail {
 
 trait ConsumerEmail extends Email
 
-object Email {
-
-  // ======= Conso =======
-
-  final case class ConsumerValidateEmail(
-      emailValidation: EmailValidation,
-      locale: Option[Locale],
-      messagesApi: MessagesApi
-  ) extends ConsumerEmail {
-    private val lang                                        = Lang(getLocaleOrDefault(locale))
-    implicit private val messagesProvider: MessagesProvider = MessagesImpl(lang, messagesApi)
-
-    override val recipients: List[EmailAddress] = List(emailValidation.email)
-    override val subject: String                = messagesApi("ConsumerValidateEmail.subject")(lang)
-
-    override def getBody: (FrontRoute, EmailAddress) => String = (frontRoute, contactAddress) =>
-      views.html.mails.consumer
-        .confirmEmail(emailValidation.email, emailValidation.confirmationCode)(
-          frontRoute,
-          contactAddress,
-          messagesProvider
-        )
-        .toString
-  }
-
-  private def getLocaleOrDefault(locale: Option[Locale]): Locale = locale.getOrElse(Locale.FRENCH)
-}
+object Email {}
